requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements.in
#
asgiref==3.5.0
    # via django
django==3.2.12
    # via
    #   -r requirements.in
    #   django-filter
    #   djangorestframework
django-environ==0.8.1
    # via -r requirements.in
django-filter==21.1
    # via -r requirements.in
django-split-settings==1.1.0
    # via -r requirements.in
djangorestframework==3.13.1
    # via -r requirements.in
djangorestframework-camel-case==1.3.0
    # via -r requirements.in
pytz==2022.1
    # via
    #   django
    #   djangorestframework
sqlparse==0.4.2
    # via django
